Rename 'internal' to 'core' and fix imports
