+++ src/commands/custom-commands/addcommand.js
@@ -1,7 +1,7 @@
-const cache = require('../../internal/cache');
+const cache = require('../../core/cache');
 const CommandBase = require('../CommandBase');
 const { Command } = require('../../database');
-const logger = require('../../internal/logger');
+const logger = require('../../core/logger');
 
 module.exports.info = {
   name: 'addcommand',
@@ -24,7 +24,7 @@ class AddCommand extends CommandBase {
 
   async process () {
     if (!this.validate()) { return; }
-    
+
     const document = await Command.findOne({ command: this.commandName }).exec();
     if (document === null) {
       await this.writeCommand();
