+++ src/commands/general/report.js
@@ -1,6 +1,6 @@
 const CommandBase = require('../CommandBase');
-const logger = require('../../internal/logger');
-const moderationLogging = require('../../internal/moderation-logging');
+const logger = require('../../core/logger');
+const moderationLogging = require('../../core/moderation-logging');
 
 module.exports.info = {
   name: 'report',
@@ -46,22 +46,22 @@ class ReportCommand extends CommandBase {
         this.channel.send('Thank you! Your report has been recorded and send to the staff of the server.');
       }
     }
-    
+
   }
 
   async getMember () {
     this.channel.send(`Please send the name and ID of the member you would like to report (Ex. Wumpus#1063), or type \`cancel\` to cancel the report. [Expires in ${this.memberOptions.time / 1000} seconds]`);
-    
+
     const collectedMessage = await this.channel.awaitMessages(this.filter, this.memberOptions);
     if (collectedMessage.size === 0) {
       this.channel.send(`Report expired (No response after ${this.memberOptions.time / 1000} seconds).`);
-      return false; 
+      return false;
     }
 
     const messageContent = collectedMessage.first().content;
     if (messageContent.toLowerCase() === 'cancel') {
       this.channel.send('Report Cancelled.');
-      return false; 
+      return false;
     }
 
     this.member = messageContent;
