+++ src/database/operations.js
@@ -1,5 +1,5 @@
 require('dotenv').config();
-const logger = require('../internal/logger');
+const logger = require('../core/logger');
 const models = require('./models');
 const mongoose = require('mongoose');
 
@@ -11,7 +11,7 @@ module.exports.connect = function () {
 module.exports.getCommands = async function () {
   const commands = await models.Command.find().exec()
     .catch(err => logger.logError(err));
-  
+
   const formattedCommandOutput = {};
   commands.forEach(command => formattedCommandOutput[command.command] = command.reply);
   return formattedCommandOutput;
@@ -20,8 +20,8 @@ module.exports.getCommands = async function () {
 module.exports.getDynamicConfigs = async function () {
   const configs = await models.DynamicConfiguration.find().exec()
     .catch(err => logger.logError(err));
-  
+
   const formattedConfigOutput = {};
   configs.forEach(config => formattedConfigOutput[config.categoryID] = config.voiceChannelName);
-  return formattedConfigOutput;  
+  return formattedConfigOutput;
 };
